+++ store/modules/data/actions.js
@@ -1,75 +1,76 @@
 // adapted from https://stackoverflow.com/questions/53446792/nuxt-vuex-how-do-i-break-down-a-vuex-module-into-separate-files
-
-import { get_recording } from "../../../store/ApiService.js";
-import {
-  convert_from_json_of_well_indices_and_x_y_arrays,
-  convert_x_y_arrays_to_d3_array,
-} from "../../../js_utils/waveform_data_formatter.js";
+import { append_well_data, find_closest_array_idx } from "../../../js_utils/waveform_data_formatter.js";
 
 export default {
-  async fetchApi({ commit }) {
-    const num_waveforms_to_display = 24;
-    const parsed_data = [];
-    try {
-      const http_response = await get_recording("sandbox_eli_waveform");
-      const json_data = http_response.data;
+  async append_plate_waveforms({ dispatch, commit, state }, new_value) {
+    const appended_waveforms = append_well_data(state.plate_waveforms, new_value);
+    commit("set_plate_waveforms", appended_waveforms);
 
-      for (let i = 0; i < num_waveforms_to_display; i++) {
-        const this_well_data = convert_from_json_of_well_indices_and_x_y_arrays(json_data, i);
-        parsed_data.push({
-          x_data_points: this_well_data.sample_indices,
-          y_data_points: this_well_data.values,
-        });
-      }
-    } catch (e) {
-      console.error(e);
-    }
-    commit("set_plate_waveforms", parsed_data);
-    return parsed_data;
+    dispatch("remove_old_waveform_data");
   },
-
+  remove_old_waveform_data({ rootState }) {
+    const { waveform, playback, data } = rootState;
+    const max_x_min_value = playback.x_time_index - waveform.x_zoom_levels[0].x_scale;
+    // remove time points indices that are older than the x min at max zoom out
+    data.plate_waveforms.map((well) => {
+      const idx_to_splice = find_closest_array_idx(well.x_data_points, max_x_min_value);
+      well.x_data_points.splice(0, idx_to_splice - 2);
+      well.y_data_points.splice(0, idx_to_splice - 2);
+    });
+  },
   async get_data_action_context(context) {
     // useful for testing actions
     return context;
   },
-
-  async append_stim_waveforms({ dispatch, state }, new_values) {
+  async append_stim_waveforms({ state, dispatch }, new_values) {
     for (const well_idx in new_values) {
       if (new_values[well_idx] !== undefined && state.plate_waveforms[well_idx] !== undefined) {
         // real Y values not actually used yet, just need to draw a straight vertical line at each new x value and connect the points at a Y value out of the max zoom window
-        const new_well_values = new_values[well_idx];
+        const previous_time = state.last_protocol_flag[well_idx];
+        const new_well_values = new_values[well_idx][0];
+        const protocol_flags = new_values[well_idx][1];
 
-        // aligns color blocking with start of tissue data when live view starts after stim
-        new_well_values[0][0] =
-          new_well_values[0][0] < state.plate_waveforms[well_idx].x_data_points[0]
-            ? state.plate_waveforms[well_idx].x_data_points[0]
-            : new_well_values[0][0];
+        if (previous_time[0]) {
+          new_well_values.unshift(previous_time[0]);
+          protocol_flags.unshift(previous_time[1]);
+        }
 
-        new_well_values[0].map((x, idx) => {
-          const y_points = [101000, -201, 101000]; // arbitrary values far enough outside of max window that the connection between vertical lines will not be rendered
-          y_points.map((y) => {
-            state.stim_waveforms[well_idx].x_data_points.push(x);
-            state.stim_waveforms[well_idx].y_data_points.push(y);
-          });
-          state.sub_protocol_flags[well_idx].push([new_well_values[1][idx], x]);
-        });
-        dispatch("assign_stim_fill_colors", well_idx);
+        let idx = 0;
+        while (idx < new_well_values.length) {
+          const x = new_well_values[idx];
+          const next_x = new_well_values[idx + 1];
+
+          state.stim_waveforms[well_idx].x_data_points.push(x);
+          state.stim_waveforms[well_idx].y_data_points.push(101000);
+
+          next_x
+            ? state.stim_fill_assignments[well_idx].push([
+                protocol_flags[idx],
+                [
+                  [x, 101000],
+                  [next_x, 101000],
+                ],
+              ])
+            : (state.last_protocol_flag[well_idx] = [x, protocol_flags[idx]]);
+
+          idx++;
+        }
       }
     }
+    dispatch("remove_old_stim_data");
   },
-  assign_stim_fill_colors({ state }, well_idx) {
-    state.stim_fill_assignments[well_idx] = []; // reset because indices are changing for each time index
+  remove_old_stim_data({ rootState }) {
+    const { waveform, playback, data } = rootState;
+    const max_x_min_value = playback.x_time_index - waveform.x_zoom_levels[0].x_scale;
 
-    const time_x_values = state.stim_waveforms[well_idx].x_data_points;
-    const time_y_values = state.stim_waveforms[well_idx].y_data_points;
+    data.stim_waveforms.map((well, well_idx) => {
+      data.stim_fill_assignments[well_idx] = data.stim_fill_assignments[well_idx].filter(
+        (x) => x[1][1][0] >= max_x_min_value
+      );
 
-    state.sub_protocol_flags[well_idx].map((sub) => {
-      const start_idx = time_x_values.indexOf(sub[1]) + 1; // want index with y-value === -201
-      const sliced_x_values = time_x_values.slice(start_idx, start_idx + 4);
-      const sliced_y_values = time_y_values.slice(start_idx, start_idx + 4);
-      const sliced_d3_subprotocol = convert_x_y_arrays_to_d3_array(sliced_x_values, sliced_y_values);
-
-      state.stim_fill_assignments[well_idx].push([sub[0], sliced_d3_subprotocol]);
+      const idx_to_splice = find_closest_array_idx(well.x_data_points, max_x_min_value);
+      well.x_data_points.splice(0, idx_to_splice - 2);
+      well.y_data_points.splice(0, idx_to_splice - 2);
     });
   },
 };